pyproject.toml
[tool.poetry]
name = "contacts"
version = "0.1.0"
description = ""
authors = ["Justas Zieminykas <justaxsz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.1.0"
pyyaml = "^6.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
